Plan 9 from Bell Labs 2006-05-22

+++ sys/src/cmd/unix/drawterm/kern/stub.c
@@ -139,7 +139,7 @@ rdb(void)
 }
 
 void
-setmalloctag(void *v, ulong tag)
+setmalloctag(void *v, uintptr tag)
 {
 	USED(v);
 	USED(tag);
